Fix install error if documentation was not built
The CMake build script provides a way to optionally build documentation.
The script was searching for the built documentation
and tried to copy it into the install location.
The script would stop with an error if the documentation was not found,
regardless of whether documentation was built.

The script is now fixed to install documentation only if it exists.

Note that, now, as long as the documentation is built even if partially,
it will always be installed.
This means it may install partially built documentation
depending on how CMake is used.
It also does not allow for the use case of building the documentation
but not installing it.